added location, static image router

+++ app/handler/post_id_delete.go
@@ -9,8 +9,14 @@ import (
 )
 
   
-func PostDeleteID(db database.DBInterface) gin.HandlerFunc{
+func PostDeleteID(db database.DBInterface, token string) gin.HandlerFunc{
 	return func(c *gin.Context){
+		auth := c.GetHeader("Authorization")
+		if auth != token {
+			c.JSON(http.StatusBadRequest, gin.H{"message": "Invalid authentication"})
+			return
+		}
+
 		id := c.Param("id")
 		res, err := db.Delete(id)
 		if err != nil {
